Clean up (#17)
* Add 127.0.0.1 for local config

* Optimize imports

* Refactor mapledataprovider

* Polishing config

* Remove old config file

* Removed docker file

* Removed maplebit for now

+++ server/src/main/java/server/maps/event/SimpleQuestMapEvent.java
@@ -6,19 +6,19 @@ import tools.MaplePacketCreator;
 
 public class SimpleQuestMapEvent extends AbstractMapEvent {
 
-  public SimpleQuestMapEvent(MapleMap map) {
-    super(map);
-  }
+    public SimpleQuestMapEvent(MapleMap map) {
+        super(map);
+    }
 
-  @Override
-  public void onUserEnter(MapleCharacter c) {
-    map.broadcastMessage(MaplePacketCreator.getClock((int) (10 * 60)));
-  }
+    @Override
+    public void onUserEnter(MapleCharacter c) {
+        map.broadcastMessage(MaplePacketCreator.getClock(10 * 60));
+    }
 
-  @Override
-  public void onUserExit(MapleCharacter c) {
-    map.killAllMonsters(true);
+    @Override
+    public void onUserExit(MapleCharacter c) {
+        map.killAllMonsters(true);
 
-  }
+    }
 
 }
